requirements.txt

cachetools
Wikipedia-API
discord.py
numpy
